backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow>=2.16.0
pillow>=10.1.0
numpy>=1.26.0
google-generativeai>=0.3.0
reportlab>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic[email]>=2.5.0
